@@ -9,38 +9,12 @@ module AsciidocBib
 
     # Valid reference styles
     def Styles.available
-      [
-        "apa",
-        "bibtex",
-        "chicago-annotated-bibliography",
-        "chicago-author-date-basque",
-        "chicago-author-date-de",
-        "chicago-author-date",
-        "chicago-dated-note-biblio-no-ibid",
-        "chicago-fullnote-bibliography-bb",
-        "chicago-fullnote-bibliography-delimiter-fixes",
-        "chicago-fullnote-bibliography-no-ibid-delimiter-fixes",
-        "chicago-fullnote-bibliography-no-ibid",
-        "chicago-fullnote-bibliography",
-        "chicago-library-list",
-        "chicago-note-biblio-no-ibid",
-        "chicago-note-bibliography",
-        "chicago-quick-copy",
-        "ieee",
-        "mla-notes",
-        "mla-underline",
-        "mla-url",
-        "mla",
-        "vancouver-brackets",
-        "vancouver-superscript-bracket-only-year",
-        "vancouver-superscript",
-        "vancouver"
-      ]
+      CSL::Styles.ls
     end
 
-    # Make the default style the first one in the list
+    # Make the default style simple APA
     def Styles.default_style
-      Styles.available.first
+      :apa
     end
 
     # Check if given style is a valid style
@@ -49,6 +23,7 @@ module AsciidocBib
     end
 
     # Test here for any numeric styles for citeproc
+    # TODO
     def Styles.is_numeric? style
       [
         "ieee",
